--- migration/src/main/kotlin/com/redbox/domain/user/dto/CheckUserRequest.kt
@@ -1,5 +1,0 @@
-package com.redbox.domain.user.dto
-
-data class CheckUserRequest(
-    val email: String
-)
