--- migration/src/main/kotlin/com/redbox/domain/user/dto/CheckUserResponse.kt
@@ -1,6 +1,0 @@
-package com.redbox.domain.user.dto
-
-data class CheckUserResponse(
-    val userId: Long,
-    val name: String
-)
