--- migration/src/main/kotlin/com/redbox/domain/user/dto/FindIdResponse.kt
@@ -1,5 +1,0 @@
-package com.redbox.domain.user.dto
-
-data class FindIdResponse(
-    val email: String
-)
