--- migration/src/main/kotlin/com/redbox/domain/user/dto/SignupResponse.kt
@@ -1,6 +1,0 @@
-package com.redbox.domain.user.dto
-
-data class SignupResponse (
-    val email: String,
-    val userName: String
-)
