--- migration/src/main/kotlin/com/redbox/domain/user/entity/Gender.kt
@@ -1,7 +1,0 @@
-package com.redbox.domain.user.entity
-
-enum class Gender(
-    val text: String
-) {
-    MALE("남성"), FEMALE("여성")
-}
