--- migration/src/main/kotlin/com/redbox/domain/user/entity/RoleType.kt
@@ -1,9 +1,0 @@
-package com.redbox.domain.user.entity
-
-enum class RoleType(
-    val text: String,
-    val fullRole: String
-) {
-    ADMIN("관리자", "ROLE_ADMIN"),
-    USER("일반 사용자", "ROLE_USER");
-}
