--- migration/src/main/kotlin/com/redbox/domain/user/entity/Status.kt
@@ -1,8 +1,0 @@
-package com.redbox.domain.user.entity
-
-enum class Status(
-    val text: String
-) {
-    ACTIVE("활성"),
-    INACTIVE("비활성")
-}
