--- migration/src/main/kotlin/com/redbox/domain/user/exception/DuplicateEmailException.kt
@@ -1,6 +1,0 @@
-package com.redbox.domain.user.exception
-
-import com.redbox.global.exception.BusinessException
-import com.redbox.global.exception.ErrorCode
-
-class DuplicateEmailException : BusinessException(ErrorCode.DUPLICATE_EMAIL)
