+++ migration/src/main/kotlin/com/redbox/domain/user/user/dto/CheckUserRequest.kt
@@ -1,0 +1,5 @@
+package com.redbox.domain.user.user.dto
+
+data class CheckUserRequest(
+    val email: String
+)
