+++ migration/src/main/kotlin/com/redbox/domain/user/user/dto/CheckUserResponse.kt
@@ -1,0 +1,6 @@
+package com.redbox.domain.user.user.dto
+
+data class CheckUserResponse(
+    val userId: Long,
+    val name: String
+)
