+++ migration/src/main/kotlin/com/redbox/domain/user/user/dto/FindIdResponse.kt
@@ -1,0 +1,5 @@
+package com.redbox.domain.user.user.dto
+
+data class FindIdResponse(
+    val email: String
+)
