+++ migration/src/main/kotlin/com/redbox/domain/user/user/dto/SignupResponse.kt
@@ -1,0 +1,6 @@
+package com.redbox.domain.user.user.dto
+
+data class SignupResponse (
+    val email: String,
+    val userName: String
+)
