+++ migration/src/main/kotlin/com/redbox/domain/user/user/dto/UpdatePasswordRequest.kt
@@ -1,0 +1,15 @@
+package com.redbox.domain.user.user.dto
+
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Pattern
+
+data class UpdatePasswordRequest(
+    @field:NotBlank(message = "비밀번호를 입력해주세요.")
+    @field:Pattern(
+        regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,}$",
+        message = "비밀번호는 8자 이상, 영문, 숫자, 특수문자를 포함해야 합니다."
+    )
+    val password: String,
+
+    val passwordConfirm: String
+)
