+++ migration/src/main/kotlin/com/redbox/domain/user/user/entity/Gender.kt
@@ -1,0 +1,7 @@
+package com.redbox.domain.user.user.entity
+
+enum class Gender(
+    val text: String
+) {
+    MALE("남성"), FEMALE("여성")
+}
