+++ migration/src/main/kotlin/com/redbox/domain/user/user/entity/RoleType.kt
@@ -1,0 +1,9 @@
+package com.redbox.domain.user.user.entity
+
+enum class RoleType(
+    val text: String,
+    val fullRole: String
+) {
+    ADMIN("관리자", "ROLE_ADMIN"),
+    USER("일반 사용자", "ROLE_USER");
+}
