+++ migration/src/main/kotlin/com/redbox/domain/user/user/entity/Status.kt
@@ -1,0 +1,8 @@
+package com.redbox.domain.user.user.entity
+
+enum class Status(
+    val text: String
+) {
+    ACTIVE("활성"),
+    INACTIVE("비활성")
+}
