+++ migration/src/main/kotlin/com/redbox/domain/user/user/exception/EmailNotVerifiedException.kt
@@ -1,0 +1,6 @@
+package com.redbox.domain.user.user.exception
+
+import com.redbox.global.exception.BusinessException
+import com.redbox.global.exception.ErrorCode
+
+class EmailNotVerifiedException: BusinessException(ErrorCode.UNVERIFIED_EMAIL)
