+++ migration/src/main/kotlin/com/redbox/domain/user/user/exception/PasswordMismatchException.kt
@@ -1,0 +1,6 @@
+package com.redbox.domain.user.user.exception
+
+import com.redbox.global.exception.AuthException
+import com.redbox.global.exception.ErrorCode
+
+class PasswordMismatchException : AuthException(ErrorCode.PASSWORD_INVALID)
