+++ migration/src/main/kotlin/com/redbox/global/auth/service/AuthenticationService.kt
@@ -2,8 +2,8 @@ package com.redbox.global.auth.service
 
 import com.redbox.domain.auth.dto.CustomUserDetails
 import com.redbox.domain.community.funding.exception.UserNotFoundException
-import com.redbox.domain.user.entity.User
-import com.redbox.domain.user.repository.UserRepository
+import com.redbox.domain.user.user.entity.User
+import com.redbox.domain.user.user.repository.UserRepository
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
